requirements.txt
pandas
wikipedia
tqdm
unidecode
faiss-cpu
sentence-transformers
streamlit